requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.0
pydantic==2.5.0

openai==1.3.0
anthropic==0.7.0

opencv-python==4.8.0
numpy==1.24.0
scipy==1.11.0
scikit-learn==1.3.0
pillow==10.0.0

face-recognition==1.3.5
dlib==19.24.2
mediapipe==0.10.0

librosa==0.10.0
soundfile==0.12.1
scipy==1.11.0
SpeechRecognition==3.10.0

torch==2.1.0
torchvision==0.16.0

sqlalchemy==2.0.0
psycopg2-binary==2.9.0

redis==5.0.0
pymongo==4.5.0

python-multipart==0.0.6
